+++ packages/core/test/instance.test-d.ts
@@ -1,9 +1,10 @@
 import { describe, expectTypeOf, test } from "vitest"
 import { createAuth } from "@/createAuth.ts"
-import { identitySchema, zod, type IdentityShape } from "@/identity/zod.ts"
+import { identitySchema, zod } from "@/identity/zod.ts"
 import type { Session, User } from "@/index.ts"
-import type { InferSignUp, Wrap } from "@/@types/utility.ts"
-import type { FromShapeToObject, Identities, InferSession, InferUser } from "@/identity/index.ts"
+import type { ZodIdentitySchema } from "@/@types/index.ts"
+import type { InferSession, InferUser } from "@/identity/index.ts"
+import type { InferSignUp, Wrap, ZodShapeToObject } from "@/@types/utility.ts"
 import type {
     GetSessionAPIReturn,
     RefreshUserInfoAPIReturn,
@@ -24,7 +25,7 @@ describe("createAuth", () => {
             },
         })
         const { api } = auth
-        type Identity = FromShapeToObject<IdentityShape & { role: zod.ZodString; nickname: zod.ZodOptional<zod.ZodString> }>
+        type Identity = ZodShapeToObject<ZodIdentitySchema & { role: zod.ZodString; nickname: zod.ZodOptional<zod.ZodString> }>
 
         test("Infer Types", () => {
             expectTypeOf<InferUser<typeof auth>>().toEqualTypeOf<Wrap<Identity>>()
@@ -54,7 +55,7 @@ describe("createAuth", () => {
         })
     })
 
-    describe("with custom identity and sign up schema", () => {
+    describe("with sign up schema", () => {
         const auth = createAuth({
             oauth: [],
             signUp: {
@@ -67,11 +68,64 @@ describe("createAuth", () => {
             },
         })
         const { api } = auth
-        type Identity = FromShapeToObject<Identities>
+        type Identity = ZodShapeToObject<ZodIdentitySchema>
 
         test("InferUser", () => {
             expectTypeOf<InferUser<typeof auth>>().toEqualTypeOf<User>()
             expectTypeOf<InferSession<typeof auth>>().toEqualTypeOf<Session<User>>()
+            expectTypeOf<InferSignUp<typeof auth>>().toEqualTypeOf<{
+                nickname: string
+                email: string
+                password: string
+            }>()
+        })
+
+        test("api.getSession", () => {
+            expectTypeOf<Awaited<ReturnType<typeof api.getSession>>>().toEqualTypeOf<GetSessionAPIReturn<Identity>>()
+        })
+
+        test("api.updateSession", () => {
+            expectTypeOf<Parameters<typeof api.updateSession>[0]>().toEqualTypeOf<UpdateSessionAPIOptions<Identity>>()
+        })
+
+        test("api.refreshUserInfo", () => {
+            expectTypeOf<ReturnType<typeof api.refreshUserInfo>>().toEqualTypeOf<Promise<RefreshUserInfoAPIReturn<Identity>>>()
+        })
+
+        test("api.signUp", () => {
+            expectTypeOf<Parameters<typeof api.signUp>[0]>().toEqualTypeOf<SignUpAPIOptions<Record<string, any>>>()
+            expectTypeOf(api.signUp).toEqualTypeOf<
+                <Payload extends Record<string, any> = InferSignUp<typeof auth>>(
+                    options: SignUpAPIOptions<Payload>
+                ) => Promise<SignUpAPIReturn>
+            >()
+        })
+    })
+
+    describe("with custom identity and sign up schema", () => {
+        const auth = createAuth({
+            oauth: [],
+            identity: {
+                schema: identitySchema.extend({
+                    role: zod.string(),
+                    nickname: zod.string().optional(),
+                }),
+            },
+            signUp: {
+                schema: zod.object({
+                    nickname: zod.string(),
+                    email: zod.email(),
+                    password: zod.string().min(8),
+                }),
+                onCreateUser: () => null,
+            },
+        })
+        const { api } = auth
+        type Identity = ZodShapeToObject<ZodIdentitySchema & { role: zod.ZodString; nickname: zod.ZodOptional<zod.ZodString> }>
+
+        test("InferUser", () => {
+            expectTypeOf<InferUser<typeof auth>>().toEqualTypeOf<Wrap<Identity>>()
+            expectTypeOf<InferSession<typeof auth>>().toEqualTypeOf<Session<Wrap<Identity>>>()
             expectTypeOf<InferSignUp<typeof auth>>().toEqualTypeOf<{
                 nickname: string
                 email: string